implemented framework for creating images via latex
Latex can now be used to created images. Tex-files in `./img/tex`
will be parsed to PDF and then split up into PNGs that can be used
within summaries. This procedure is automated via `make`. To tell
the makefile that the compilation of a certain tex-file in
`./img/tex` is required, add `tikz-requirements: my-tex-file-basename`
to the yaml metadata header. (here, tikz is just representative
for all images created with latex).